Add code docs for the ManagedObjectController scopes feature
