+++ aqueduct/lib/src/http/managed_object_controller_action_scopes.dart
@@ -1,0 +1,13 @@
+/// Allows [ManagedObjectController]s to have different scopes for each CRUD action.
+///
+/// An instance of it could be passed to [ManagedObjectController] constructor to enable that behavior
+class ActionScopes {
+  const ActionScopes(
+      {this.index, this.delete, this.update, this.create, this.find});
+
+  final List<String> find;
+  final List<String> index;
+  final List<String> create;
+  final List<String> update;
+  final List<String> delete;
+}
